requirements.txt
torchmetrics==1.5.2
torchsampler==0.1.2
accelerate==1.0.1
albumentations==1.4.18
numpy==1.24.4
opencv-python==4.8.0.74
Pillow==9.5.0
requests==2.32.3
tqdm==4.67.0
wandb==0.19.2
einops==0.8.0
scipy==1.10.1
matplotlib==3.7.1
kornia==0.7.3